pyproject.toml
[tool.poetry]
name = "datalakehouse-lowcost"
version = "0.1.0"
description = ""
authors = ["marcelo-borges <marcelo.b.orges@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.3"
pandas = "^2.1.4"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pydantic = "^2.5.3"


[tool.poetry.group.dev.dependencies]
commitizen = "^3.13.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
